third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-invalid-syntax.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    --green: {
        background-color: green;
    };
    @apply --green;

    @apply --unknown;
    --red: {
        background-color: red;
    };
    @apply--red;
    @apply red;
    @apply(--red);
    @apply --red !important;
    @apply --red { }
    @apply { background-color: red; }
    --redvar: --red;
    @apply var(--redvar);

    --red2: { background-color: red; } bla ;
    @apply --red2;

    width: 100px;
    height: 100px;

    animation: anim 0s both;
}

@apply --red;

@keyframes anim {
    100% {
        @apply --red;
    }
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
}, "Tests various invalid @apply rules get rejected");
</script>
